[Feat] 수정 화면 구현

+++ core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/vo/memo/MemoEditParameter.kt
@@ -6,6 +6,6 @@ data class MemoEditParameter(
     val title: String?,
     val description: String?,
     val isCompleted: Boolean?,
-    val tagIds: List<Int>?,
+    val tagIds: List<Long>?,
     val dateTimeInfo: DateTimeInfo?
 )
